scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = ""
authors = [
    {name = "moizahmedshaikh",email = "moizaman905@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastembed (>=0.7.3,<0.8.0)",
    "qdrant-client (>=1.16.1,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
